move the class methods for starting watchers to the watcher class
