+++ plugins/kubernetes/app/models/kubernetes/util.rb
@@ -1,5 +1,3 @@
-require 'celluloid/current'
-
 module Kubernetes
   module Util
     def self.parse_file(contents, filepath)
@@ -21,21 +19,6 @@ module Kubernetes
       @output_file.write({ message: message }.merge(extra_info).to_json + "\n")
       @output_file.flush
       message
-    end
-
-    def self.pod_watcher_symbol(cluster)
-      "cluster_pod_watcher_#{cluster.id}".to_sym
-    end
-
-    def self.start_watcher(cluster)
-      watcher_name = pod_watcher_symbol(cluster)
-      Watchers::ClusterPodWatcher.supervise as: watcher_name, args: [cluster.client]
-    end
-
-    def self.restart_watcher(cluster)
-      watcher = Celluloid::Actor[pod_watcher_symbol(cluster)]
-      watcher.terminate if watcher and watcher.alive?
-      start_watcher(cluster)
     end
   end
 end
